LPS-63272 Rename packages
