+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/ExpressionFactoryImpl.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.expression.internal;
+package com.liferay.dynamic.data.mapping.expression.internal;
 
-import com.liferay.portal.expression.Expression;
-import com.liferay.portal.expression.ExpressionFactory;
+import com.liferay.dynamic.data.mapping.expression.Expression;
+import com.liferay.dynamic.data.mapping.expression.ExpressionFactory;
 
 import org.osgi.service.component.annotations.Component;
