+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/ExpressionImpl.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.expression.internal;
+package com.liferay.dynamic.data.mapping.expression.internal;
 
-import com.liferay.portal.expression.Expression;
-import com.liferay.portal.expression.ExpressionEvaluationException;
-import com.liferay.portal.expression.VariableDependencies;
+import com.liferay.dynamic.data.mapping.expression.Expression;
+import com.liferay.dynamic.data.mapping.expression.ExpressionEvaluationException;
+import com.liferay.dynamic.data.mapping.expression.VariableDependencies;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
