+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/VariableNamesExtractor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.expression.internal;
+package com.liferay.dynamic.data.mapping.expression.internal;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
